backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
python-multipart==0.0.6
boto3==1.34.0
requests==2.31.0
PyPDF2==3.0.1
python-docx==1.1.0
nltk==3.8.1
websockets==12.0
aiofiles==23.2.1